+++ metadata.rb
@@ -1,11 +1,11 @@
-name             'blockdevice-nativex'
-maintainer       'NativeX'
-maintainer_email 'adrian.herrera@nativex.com'
+name             'blockdevice-simplyadrian'
+maintainer       'simplyadrian'
+maintainer_email 'simplyadrian@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures blockdevice-nativex'
+description      'Installs/Configures blockdevice-simplyadrian'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.3.0'
 
 depends  "aws"
 depends  "xfs"
-depends  "ohai-nativex"
+depends  "ohai-simplyadrian"