+++ recipes/default.rb
@@ -1,8 +1,8 @@
 #
-# Cookbook Name:: blockdevice-nativex
+# Cookbook Name:: blockdevice-simplyadrian
 # Recipe:: default
 #
-# Copyright 2014, NativeX
+# Copyright 2014, simplyadrian
 #
 # All rights reserved - Do Not Redistribute
 #
@@ -11,5 +11,5 @@ gem_package "aws-sdk-v1" do
   action :install
 end
 
-include_recipe "xfs::default" if node['blockdevice_nativex']['filesystem'] == "xfs"
-include_recipe "blockdevice-nativex::volumes"
+include_recipe "xfs::default" if node['blockdevice_simplyadrian']['filesystem'] == "xfs"
+include_recipe "blockdevice-simplyadrian::volumes"