+++ recipes/volumes.rb
@@ -1,15 +1,15 @@
-directory node['blockdevice_nativex']['dir'] do
-  group node['blockdevice_nativex']['mount_point_group']
+directory node['blockdevice_simplyadrian']['dir'] do
+  group node['blockdevice_simplyadrian']['mount_point_group']
   mode 775
   recursive true
   action :create
-  not_if { ::File.directory?("#{node['blockdevice_nativex']['dir']}") }
+  not_if { ::File.directory?("#{node['blockdevice_simplyadrian']['dir']}") }
 end
 
-if node['blockdevice_nativex']['ec2'] || node['cloud']['provider'] == 'ec2'
+if node['blockdevice_simplyadrian']['ec2'] || node['cloud']['provider'] == 'ec2'
   aws = Chef::EncryptedDataBagItem.load("credentials", "aws")
   include_recipe 'aws'
-  ::Chef::Recipe.send(:include, Nativex::Blockdevice::Helpers)
+  ::Chef::Recipe.send(:include, simplyadrian::Blockdevice::Helpers)
 
   # Determine if this is a HVM or Paravirtual instance
   instance_id = get_instance_id
@@ -19,19 +19,19 @@ if node['blockdevice_nativex']['ec2'] || node['cloud']['provider'] == 'ec2'
   elsif virtualization_type == :paravirtual
     hvm = false
   else
-    hvm = node['blockdevice_nativex']['ebs']['hvm']
+    hvm = node['blockdevice_simplyadrian']['ebs']['hvm']
   end
 
-  if node['blockdevice_nativex']['ebs']['raid']
+  if node['blockdevice_simplyadrian']['ebs']['raid']
 
     aws_ebs_raid 'data_volume_raid' do
-      mount_point node['blockdevice_nativex']['dir']
-      mount_point_group node['blockdevice_nativex']['mount_point_group']
-      disk_count node['blockdevice_nativex']['ebs']['count']
-      disk_size node['blockdevice_nativex']['ebs']['size']
+      mount_point node['blockdevice_simplyadrian']['dir']
+      mount_point_group node['blockdevice_simplyadrian']['mount_point_group']
+      disk_count node['blockdevice_simplyadrian']['ebs']['count']
+      disk_size node['blockdevice_simplyadrian']['ebs']['size']
       hvm hvm
-      level node['blockdevice_nativex']['ebs']['level']
-      filesystem node['blockdevice_nativex']['filesystem']
+      level node['blockdevice_simplyadrian']['ebs']['level']
+      filesystem node['blockdevice_simplyadrian']['filesystem']
       action :auto_attach
     end
 
@@ -61,9 +61,9 @@ if node['blockdevice_nativex']['ec2'] || node['cloud']['provider'] == 'ec2'
       aws_ebs_volume 'data_volume' do
         aws_access_key aws['aws_access_key_id']
         aws_secret_access_key aws['aws_secret_access_key']
-        size node['blockdevice_nativex']['ebs']['size']
+        size node['blockdevice_simplyadrian']['ebs']['size']
         device (hvm ? device_id : device_id.gsub('xvd', 'sd')) # aws uses sdx instead of xvdx
-        most_recent_snapshot node['blockdevice_nativex']['ebs']['most_recent']
+        most_recent_snapshot node['blockdevice_simplyadrian']['ebs']['most_recent']
         action [:create, :attach]
       end
  
@@ -71,7 +71,7 @@ if node['blockdevice_nativex']['ec2'] || node['cloud']['provider'] == 'ec2'
       ruby_block "sleeping_data_volume" do
         block do
           timeout = 0
-          until File.blockdev?(device_id) || timeout >= node['blockdevice_nativex']['max_timeout']
+          until File.blockdev?(device_id) || timeout >= node['blockdevice_simplyadrian']['max_timeout']
             Chef::Log.debug("device #{device_id} not ready - sleeping 10s")
             timeout += 10
             sleep 10
@@ -81,17 +81,17 @@ if node['blockdevice_nativex']['ec2'] || node['cloud']['provider'] == 'ec2'
 
       # create a filesystem
       execute 'mkfs' do
-        command "mkfs -t #{node['blockdevice_nativex']['filesystem']} #{device_id}"
+        command "mkfs -t #{node['blockdevice_simplyadrian']['filesystem']} #{device_id}"
 
         # Note the escaped quotes for bash
         # blkid works on CentOS and hopefully elsewhere. See: http://unix.stackexchange.com/a/53552/55079
-        # TYPE=\\\"#{node['blockdevice_nativex']['filesystem']}\\\" seems to work for 'xfs' . If it doesn't work for something else, we might want a mapping of mkfs -t arguments to blkid outputs.
-        not_if "blkid #{device_id} | grep \" TYPE=\\\"#{node['blockdevice_nativex']['filesystem']}\\\"\""
+        # TYPE=\\\"#{node['blockdevice_simplyadrian']['filesystem']}\\\" seems to work for 'xfs' . If it doesn't work for something else, we might want a mapping of mkfs -t arguments to blkid outputs.
+        not_if "blkid #{device_id} | grep \" TYPE=\\\"#{node['blockdevice_simplyadrian']['filesystem']}\\\"\""
       end
 
-      mount node['blockdevice_nativex']['dir'] do
+      mount node['blockdevice_simplyadrian']['dir'] do
         device device_id
-        fstype node['blockdevice_nativex']['filesystem']
+        fstype node['blockdevice_simplyadrian']['filesystem']
         options 'noatime'
         action [:mount]
       end
@@ -100,15 +100,15 @@ if node['blockdevice_nativex']['ec2'] || node['cloud']['provider'] == 'ec2'
 
   permission_recurse_switch = 'R'
 
-  permission_recurse_switch = '' unless node['blockdevice_nativex']['recurse_permissions']
+  permission_recurse_switch = '' unless node['blockdevice_simplyadrian']['recurse_permissions']
 
-  execute "fixup #{node['blockdevice_nativex']['dir']} group" do
-    command "chown -#{permission_recurse_switch}f :#{node['blockdevice_nativex']['mount_point_group']} #{node['blockdevice_nativex']['dir']}"
-    only_if { Etc.getgrgid(File.stat("#{node['blockdevice_nativex']['dir']}").gid).name != "#{node['blockdevice_nativex']['mount_point_group']}" }
+  execute "fixup #{node['blockdevice_simplyadrian']['dir']} group" do
+    command "chown -#{permission_recurse_switch}f :#{node['blockdevice_simplyadrian']['mount_point_group']} #{node['blockdevice_simplyadrian']['dir']}"
+    only_if { Etc.getgrgid(File.stat("#{node['blockdevice_simplyadrian']['dir']}").gid).name != "#{node['blockdevice_simplyadrian']['mount_point_group']}" }
     ignore_failure true
   end
 
-  execute "fixup #{node['blockdevice_nativex']['dir']} permissions" do
-    command "chmod -#{permission_recurse_switch}f 775 #{node['blockdevice_nativex']['dir']}"
+  execute "fixup #{node['blockdevice_simplyadrian']['dir']} permissions" do
+    command "chmod -#{permission_recurse_switch}f 775 #{node['blockdevice_simplyadrian']['dir']}"
   end
 end